requirements.txt
pandas
numpy
python-dotenv
openai
markdown
matplotlib
tabulate
